requirements.txt
gym==0.26.2
numpy==1.26.4
opencv_python==4.9.0.80
opendatasets==0.1.22
Pillow==10.3.0
pycuda==2024.1
skimage==0.0
torch==2.3.1
tqdm==4.66.2
wandb==0.17.1
